Move keywords to their own header file.
We need to refactor to make FieldData accessible from the simulator
for running in parallel with the TranCalculator.
